+++ web/controllers/admin/image_category_controller.ex
@@ -114,7 +114,7 @@ defmodule Brando.Admin.ImageCategoryController do
       {:ok, _} ->
         conn
         |> put_flash(:notice, gettext("Configuration updated"))
-        |> redirect(to: helpers(conn).admin_image_path(conn, :index))
+        |> redirect(to: helpers(conn).admin_image_category_path(conn, :configure, id))
       {:error, changeset} ->
         conn
         |> assign(:page_title, gettext("Configure image category"))
@@ -152,7 +152,7 @@ defmodule Brando.Admin.ImageCategoryController do
 
     conn
     |> put_flash(:notice, gettext("Category propagated"))
-    |> redirect(to: helpers(conn).admin_image_path(conn, :index))
+    |> redirect(to: helpers(conn).admin_image_category_path(conn, :configure, id))
   end
 
   @doc false