+++ web/controllers/admin/image_series_controller.ex
@@ -119,7 +119,7 @@ defmodule Brando.Admin.ImageSeriesController do
       {:ok, _} ->
         conn
         |> put_flash(:notice, gettext("Configuration updated"))
-        |> redirect(to: helpers(conn).admin_image_path(conn, :index))
+        |> redirect(to: helpers(conn).admin_image_series_path(conn, :configure, id))
       {:error, changeset} ->
         conn
         |> assign(:page_title, gettext("Configure image series"))